Move the namespace and remove backoff implementations
* Move the namespace from api.retries to retries.api to make it
consistent with the identity module
* Remove the backoff implementations and static methods from the
interface, those classes will be living in the implementation package.
