+++ core/retries-api/src/main/java/software/amazon/awssdk/retries/api/RefreshRetryTokenResponse.java
@@ -13,18 +13,18 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.api.retries;
+package software.amazon.awssdk.retries.api;
 
 import java.time.Duration;
 import software.amazon.awssdk.annotations.SdkPublicApi;
 import software.amazon.awssdk.annotations.ThreadSafe;
 
 /**
- * Encapsulates the response from the {@link RetryStrategy} to the request to start the attempts to be executed.
+ * Response from the {@link RetryStrategy} after calling {@link RetryStrategy#refreshRetryToken(RefreshRetryTokenRequest)}.
  */
 @SdkPublicApi
 @ThreadSafe
-public interface AcquireInitialTokenResponse {
+public interface RefreshRetryTokenResponse {
     /**
      * A {@link RetryToken} acquired by this invocation, used in subsequent {@link RetryStrategy#refreshRetryToken} or
      * {@link RetryStrategy#recordSuccess} calls.
@@ -32,7 +32,7 @@ public interface AcquireInitialTokenResponse {
     RetryToken token();
 
     /**
-     * The amount of time to wait before performing the first attempt.
+     * The amount of time to wait before performing the next attempt.
      */
     Duration delay();
 }
