+++ core/retries-api/src/main/java/software/amazon/awssdk/retries/api/RetryStrategy.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.api.retries;
+package software.amazon.awssdk.retries.api;
 
 import java.util.function.Predicate;
 import software.amazon.awssdk.annotations.SdkPublicApi;
